Add first middleware
