+++ lib/api/errorCodes.js
@@ -1,6 +1,7 @@
 const errorMessages = {
   login: {
     400: {
+      format: { code: 400, message: "transmitted data corrupted" },
       id: { code: 400, message: "UserID is missing" },
       pwd: { code: 400, message: "password is missing" },
     },
