+++ lib/routes/login.js
@@ -1,24 +1,17 @@
 const { Router } = require("express");
 const { findOne } = require("../api/database");
 const { errorMessages } = require("../api/errorCodes");
+const isValidHeader = require("../api/isValidHeader");
 
 const router = Router();
 
-router.get("/login", async (request, response) => {
+router.get("/login", isValidHeader, async (request, response) => {
   try {
-    const authorizationHeader = request.headers.authorization;
-    const buff = Buffer.from(authorizationHeader.split(" ")[1], "base64");
-    const { id, password } = JSON.parse(buff.toString("utf-8"));
-
-    if (!id) {
-      response.status(400).json(errorMessages.login[400].id);
-      return;
-    }
-
-    if (!password) {
-      response.status(400).json(errorMessages.login[400].pwd);
-      return;
-    }
+    const buff = Buffer.from(
+      request.headers.authorization.split(" ")[1],
+      "base64"
+    );
+    const [id, password] = buff.toString("utf-8").split(":");
 
     const result = await findOne(process.env.DB_COLLECTION_USER, {
       $or: [{ employeeID: id }, { name: id }, { email: id }],
@@ -30,7 +23,7 @@ router.get("/login", async (request, response) => {
     }
 
     if (result.password !== password) {
-      response.status(401).json(errorMessages.login[501]);
+      response.status(401).json(errorMessages.login[401]);
       return;
     }
