+++ lib/routes/overviewDay.js
@@ -7,7 +7,6 @@ const { createPeriodOfTime } = require("../utils/createPeriodOfTime");
 const router = Router();
 
 router.get("/:timestamp", async (request, response) => {
-  console.log("route evoked");
   try {
     if (
       request.headers.authorization.split(" ")[1] !== process.env.AUTH_TOKEN
